Fixed Alpha issue
Fixed transparency not working right and put the folder into a full
Unity project.
