+++ Assets/Turbulence-Library/Editor/NoiseStrings2D.cs
