+++ Assets/Turbulence-Library/LICENSE.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 709acb81e88c74a79a29eda4bd84ebde
+timeCreated: 1488905214
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
