+++ Assets/Turbulence-Library/Material.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 744aebef4282c4e9b9d644a4c78257b4
+folderAsset: yes
+timeCreated: 1488905921
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
