+++ Assets/Turbulence-Library/Material/TestMaterial.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 691d1055abf0d42e38df0d72ccece7fe
+timeCreated: 1488905942
+licenseType: Pro
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
