+++ Assets/Turbulence-Library/Resources/shader.template
