+++ Assets/Turbulence-Library/Shaders.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 663bac9a953364b0cbb44c02732015c2
+folderAsset: yes
+timeCreated: 1488905928
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
