+++ Assets/Turbulence-Library/Shaders/3DRidged Perlin.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d4194e1a0d1f844dabe91018cc913188
+timeCreated: 1488906533
+licenseType: Pro
+ShaderImporter:
+  defaultTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
